Switched from env variables to env-file
Docker has the ability to load environment variables from a file rather
than having them specified one-by-one on the command line. (This feature
is not documented on their website.)

This commit switches from using a YAML file, which was in turn processed
by Vagrant and turned into command line options, to a regular .env file
that is passed directly to Docker.