Added missing SendWithUs Apache environment
SendWithUs key was added to the Docker image, and the config file was
added to the application, but the key was not pushed into the Apache
server environment via the vhost config file.

Changed apache-vhost.conf to add the SendWithUs key into the $_SERVER
environment.

Follows-Up: 6c936fba8a2012e1ec1f74e31bd5911f79974863