Handle upgrade to Debian Jessie and PHP 5.5
Cleaned up Dockerfile and Makefile, moving Composer installation to the
Makefile so that it does not need to be installed into the Docker
image. Also removed the SSH key, as it is not really necessary.

Finished adjustment of packages for Debian Jessie, switching from
mod_fastcgi to mod_proxy_fcgi, and replacing the PHP5-ElastiCache
extension with a version compiled for PHP 5.5.

Finally, added the dotdeb repository to allow for downgrading PHP to 5.5
rather than the 5.6 default that comes with Jessie.